# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attendance Management API for Police Training Units
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UsuarioRol = (typeof UsuarioRol)[keyof typeof UsuarioRol];

export const UsuarioRol = {
  superusuario: "superusuario",
  estandar: "estandar",
} as const;

export interface Usuario {
  id: number;
  email: string;
  nombre: string;
  rol: UsuarioRol;
  pelotonId?: number | null;
  pelotonNombre?: string | null;
  activo: boolean;
  createdAt: string;
}

export interface LoginResponse {
  token: string;
  usuario: Usuario;
}

export interface Proceso {
  id: number;
  nombre: string;
  activo: boolean;
  fechaArchivado?: string | null;
  createdAt: string;
}

export interface CreateProcesoRequest {
  nombre: string;
}

export interface UpdateProcesoRequest {
  nombre?: string;
  activo?: boolean;
}

export interface Pnf {
  id: number;
  nombre: string;
  createdAt: string;
}

export interface CreatePnfRequest {
  nombre: string;
}

export interface UpdatePnfRequest {
  nombre?: string;
}

export interface Peloton {
  id: number;
  nombre: string;
  procesoId: number;
  procesoNombre: string;
  pnfId: number;
  pnfNombre: string;
  totalPersonas: number;
  createdAt: string;
}

export interface CreatePelotonRequest {
  nombre: string;
  procesoId: number;
  pnfId: number;
}

export interface UpdatePelotonRequest {
  nombre?: string;
  procesoId?: number;
  pnfId?: number;
}

export type PersonaSexo = (typeof PersonaSexo)[keyof typeof PersonaSexo];

export const PersonaSexo = {
  M: "M",
  F: "F",
} as const;

export interface Persona {
  id: number;
  nombres: string;
  apellidos: string;
  ci: string;
  sexo: PersonaSexo;
  pelotonId: number;
  pelotonNombre: string;
  createdAt: string;
  updatedAt: string;
}

export type CreatePersonaRequestSexo =
  (typeof CreatePersonaRequestSexo)[keyof typeof CreatePersonaRequestSexo];

export const CreatePersonaRequestSexo = {
  M: "M",
  F: "F",
} as const;

export interface CreatePersonaRequest {
  nombres: string;
  apellidos: string;
  ci: string;
  sexo: CreatePersonaRequestSexo;
  pelotonId: number;
}

export type UpdatePersonaRequestSexo =
  (typeof UpdatePersonaRequestSexo)[keyof typeof UpdatePersonaRequestSexo];

export const UpdatePersonaRequestSexo = {
  M: "M",
  F: "F",
} as const;

export interface UpdatePersonaRequest {
  nombres?: string;
  apellidos?: string;
  ci?: string;
  sexo?: UpdatePersonaRequestSexo;
  pelotonId?: number;
}

export interface PlanBusqueda {
  id: number;
  personaId: number;
  telefono1?: string | null;
  telefono2?: string | null;
  telefono3?: string | null;
  direccion?: string | null;
  lugarOrigen?: string | null;
}

export interface UpdatePlanBusquedaRequest {
  telefono1?: string | null;
  telefono2?: string | null;
  telefono3?: string | null;
  direccion?: string | null;
  lugarOrigen?: string | null;
}

export type AsistenciaEstado =
  (typeof AsistenciaEstado)[keyof typeof AsistenciaEstado];

export const AsistenciaEstado = {
  inasistente: "inasistente",
  presente: "presente",
  comision: "comision",
  reposo: "reposo",
} as const;

export interface Asistencia {
  id: number;
  personaId: number;
  personaNombres: string;
  personaApellidos: string;
  personaCi: string;
  personaSexo: string;
  pelotonId: number;
  fecha: string;
  estado: AsistenciaEstado;
  motivo?: string | null;
  usuarioId: number;
  createdAt: string;
  updatedAt: string;
}

export type AsistenciaRecordEstado =
  (typeof AsistenciaRecordEstado)[keyof typeof AsistenciaRecordEstado];

export const AsistenciaRecordEstado = {
  inasistente: "inasistente",
  presente: "presente",
  comision: "comision",
  reposo: "reposo",
} as const;

export interface AsistenciaRecord {
  personaId: number;
  estado: AsistenciaRecordEstado;
  motivo?: string | null;
}

export interface SaveAsistenciasRequest {
  pelotonId: number;
  fecha: string;
  registros: AsistenciaRecord[];
}

export interface PelotonStats {
  pelotonId: number;
  pelotonNombre: string;
  pnfNombre: string;
  procesoNombre: string;
  total: number;
  presentes: number;
  presentesH: number;
  presentesM: number;
  inasistentes: number;
  inasistentesH: number;
  inasistentesM: number;
  comisiones: number;
  comisionesH: number;
  comisionesM: number;
  reposos: number;
  reposesH: number;
  reposesM: number;
}

export interface InasistentePerson {
  personaId: number;
  nombres: string;
  apellidos: string;
  ci: string;
  sexo: string;
  pelotonNombre: string;
  pnfNombre: string;
  procesoNombre: string;
  estado: string;
  motivo?: string | null;
}

export type CreateUsuarioRequestRol =
  (typeof CreateUsuarioRequestRol)[keyof typeof CreateUsuarioRequestRol];

export const CreateUsuarioRequestRol = {
  superusuario: "superusuario",
  estandar: "estandar",
} as const;

export interface CreateUsuarioRequest {
  email: string;
  nombre: string;
  password: string;
  rol: CreateUsuarioRequestRol;
  pelotonId?: number | null;
}

export type UpdateUsuarioRequestRol =
  (typeof UpdateUsuarioRequestRol)[keyof typeof UpdateUsuarioRequestRol];

export const UpdateUsuarioRequestRol = {
  superusuario: "superusuario",
  estandar: "estandar",
} as const;

export interface UpdateUsuarioRequest {
  email?: string;
  nombre?: string;
  password?: string;
  rol?: UpdateUsuarioRequestRol;
  pelotonId?: number | null;
  activo?: boolean;
}

export type GetProcesosParams = {
  includeArchived?: boolean;
};

export type GetPelotonesParams = {
  procesoId?: number;
  pnfId?: number;
};

export type GetPersonasParams = {
  pelotonId?: number;
  search?: string;
};

export type GetAsistenciasParams = {
  pelotonId?: number;
  fecha?: string;
  procesoId?: number;
  pnfId?: number;
  estado?: string;
  sexo?: string;
};

export type GetDashboardParams = {
  fecha: string;
  procesoId?: number;
  pnfId?: number;
};

export type GetInasistentesParams = {
  fecha: string;
  procesoId?: number;
  pnfId?: number;
  pelotonId?: number;
};