# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Attendance Management API for Police Training Units
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /pelotones:
    get:
      operationId: getPelotones
      tags: [pelotones]
      summary: List squads
      parameters:
        - name: procesoId
          in: query
          schema:
            type: integer
        - name: pnfId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of squads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Peloton"
    post:
      operationId: createPeloton
      tags: [pelotones]
      summary: Create squad
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePelotonRequest"
      responses:
        "201":
          description: Created squad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Peloton"
  /pelotones/{id}:
    put:
      operationId: updatePeloton
      tags: [pelotones]
      summary: Update squad
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePelotonRequest"
      responses:
        "200":
          description: Updated squad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Peloton"
    delete:
      operationId: deletePeloton
      tags: [pelotones]
      summary: Delete squad
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /asistencias:
    get:
      operationId: getAsistencias
      tags: [asistencias]
      summary: Get attendance records
      parameters:
        - name: pelotonId
          in: query
          schema:
            type: integer
        - name: fecha
          in: query
          schema:
            type: string
        - name: procesoId
          in: query
          schema:
            type: integer
        - name: pnfId
          in: query
          schema:
            type: integer
        - name: estado
          in: query
          schema:
            type: string
        - name: sexo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asistencia"
    post:
      operationId: saveAsistencias
      tags: [asistencias]
      summary: Save attendance for a squad and date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveAsistenciasRequest"
      responses:
        "200":
          description: Saved attendance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /asistencias/dashboard:
    get:
      operationId: getDashboard
      tags: [asistencias]
      summary: Get dashboard stats
      parameters:
        - name: fecha
          in: query
          required: true
          schema:
            type: string
        - name: procesoId
          in: query
          schema:
            type: integer
        - name: pnfId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PelotonStats"
  /asistencias/inasistentes:
    get:
      operationId: getInasistentes
      tags: [asistencias]
      summary: Get all absentees for export
      parameters:
        - name: fecha
          in: query
          required: true
          schema:
            type: string
        - name: procesoId
          in: query
          schema:
            type: integer
        - name: pnfId
          in: query
          schema:
            type: integer
        - name: pelotonId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of absentees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InasistentePerson"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        usuario:
          $ref: "#/components/schemas/Usuario"
      required: [token, usuario]
    Usuario:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        nombre:
          type: string
        rol:
          type: string
          enum: [superusuario, estandar]
        pelotonId:
          type: integer
          nullable: true
        pelotonNombre:
          type: string
          nullable: true
        activo:
          type: boolean
        createdAt:
          type: string
      required: [id, email, nombre, rol, activo, createdAt]
...
    Peloton:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        procesoId:
          type: integer
        procesoNombre:
          type: string
        pnfId:
          type: integer
        pnfNombre:
          type: string
        totalPersonas:
          type: integer
        createdAt:
          type: string
      required: [id, nombre, procesoId, procesoNombre, pnfId, pnfNombre, totalPersonas, createdAt]
    CreatePelotonRequest:
      type: object
      properties:
        nombre:
          type: string
        procesoId:
          type: integer
        pnfId:
          type: integer
      required: [nombre, procesoId, pnfId]
    UpdatePelotonRequest:
      type: object
      properties:
        nombre:
          type: string
        procesoId:
          type: integer
        pnfId:
          type: integer
...
    Asistencia:
      type: object
      properties:
        id:
          type: integer
        personaId:
          type: integer
        personaNombres:
          type: string
        personaApellidos:
          type: string
        personaCi:
          type: string
        personaSexo:
          type: string
        pelotonId:
          type: integer
        fecha:
          type: string
        estado:
          type: string
          enum: [inasistente, presente, comision, reposo]
        motivo:
          type: string
          nullable: true
        usuarioId:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, personaId, personaNombres, personaApellidos, personaCi, personaSexo, pelotonId, fecha, estado, usuarioId, createdAt, updatedAt]
    AsistenciaRecord:
      type: object
      properties:
        personaId:
          type: integer
        estado:
          type: string
          enum: [inasistente, presente, comision, reposo]
        motivo:
          type: string
          nullable: true
      required: [personaId, estado]
    SaveAsistenciasRequest:
      type: object
      properties:
        pelotonId:
          type: integer
        fecha:
          type: string
        registros:
          type: array
          items:
            $ref: "#/components/schemas/AsistenciaRecord"
      required: [pelotonId, fecha, registros]
    PelotonStats:
      type: object
      properties:
        pelotonId:
          type: integer
        pelotonNombre:
          type: string
        pnfNombre:
          type: string
        procesoNombre:
          type: string
        total:
          type: integer
        presentes:
          type: integer
        presentesH:
          type: integer
        presentesM:
          type: integer
        inasistentes:
          type: integer
        inasistentesH:
          type: integer
        inasistentesM:
          type: integer
        comisiones:
          type: integer
        comisionesH:
          type: integer
        comisionesM:
          type: integer
        reposos:
          type: integer
        reposesH:
          type: integer
        reposesM:
          type: integer
      required: [pelotonId, pelotonNombre, pnfNombre, procesoNombre, total, presentes, presentesH, presentesM, inasistentes, inasistentesH, inasistentesM, comisiones, comisionesH, comisionesM, reposos, reposesH, reposesM]
    InasistentePerson:
      type: object
      properties:
        personaId:
          type: integer
        nombres:
          type: string
        apellidos:
          type: string
        ci:
          type: string
        sexo:
          type: string
        pelotonNombre:
          type: string
        pnfNombre:
          type: string
        procesoNombre:
          type: string
        estado:
          type: string
        motivo:
          type: string
          nullable: true
      required: [personaId, nombres, apellidos, ci, sexo, pelotonNombre, pnfNombre, procesoNombre, estado]
...