# lib/api-zod/src/generated/types/asistenciaRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attendance Management API for Police Training Units
 * OpenAPI spec version: 0.1.0
 */
import type { AsistenciaRecordEstado } from "./asistenciaRecordEstado";

export interface AsistenciaRecord {
  personaId: number;
  estado: AsistenciaRecordEstado;
  motivo?: string | null;
}