# lib/api-zod/src/generated/types/createPelotonRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attendance Management API for Police Training Units
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePelotonRequest {
  nombre: string;
  procesoId: number;
  pnfId: number;
}