# lib/api-zod/src/generated/types/peloton.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attendance Management API for Police Training Units
 * OpenAPI spec version: 0.1.0
 */

export interface Peloton {
  id: number;
  nombre: string;
  procesoId: number;
  procesoNombre: string;
  pnfId: number;
  pnfNombre: string;
  totalPersonas: number;
  createdAt: string;
}